fix: Moved fixture to pytest file itself, try if avoid CI/CD test failed(Success in local)
